Clean up unnecessary codepaths from operator-run
I'm investigating a bug from #9005, and discovered that many of these codepaths are rarely or never exercised by our tests, because there are just too many of them (and some of them are simply dead code). The probability of groups = 1 and batches = 1 in a randomized test is low, and when running on slow emulators, we don't get many chances. We also have different code for ARM vs. not-ARM (via `XNN_MAX_UARCH_TYPES`), which again forks our test coverage.

I don't think these specializations are worth the cost (the constant vigilance required to ensure we don't lose test coverage of all of these paths).

PiperOrigin-RevId: 826767227